feat: sort by name and column
This step of the exercise was very troubling while working on it in the
heat of the moment. My first thought while doing it was to use the
`prevState` prop of the callback to avoid any cyclic depedencies with
`filteredData` (since it would be changed by the effect and then it
would retrigger it). However, I was biased by the continuous usage of
ESLint on the projects I worked were dependencies need to be put there,
always.

In the framework of this exercise, such check was not needed so the
easiest way to develop it was to change `filteredData`.

The approach I used was mutating the `prevState` element in place (by
the function sort). So this would not trigger a re-render on the page
which would resort in an unexpected UX.

The final workaround was to destruct the mutated element inside a new
array, effectively copying it and thus forcing a re-render of the page.