feat: filter by support of testmode
Initially, I thought about using a Pressable component to toggle states.
This would have been good in case we need to create our own toggling
component. However, as said by Iacopo, we didn't need that UI refinement
so I went to the ReactNative docs to pull the sample code for the Switch
component and use that instead. A few minor modifications were needed
since code was different but it was basically it.

Wanted to make a row view to put the text alongside it but during the
exercise it was expressed the lack of need of that so it was left as a
column container which doesn't make much sense visually but works for
the test.