feat: add test to assert if filtering is working properly
The easiest way, although flawed, was to check the screen for any
`false` texts since they should have been filtered and not be rendered.

This is a bad test since it could happen that an element of the list has
either a name or code "false" so it would trigger a false negative.

Discussing with the team, they asked me if I was familiar with
`@testing-library` (to which I replied I used just a few times) because
there is a built-in function which is better suited for this.

After a quick google search while I was been told that it was no longer
needed I recalled that there's a queryByText function which would have
been better suited to test the lack of element in the screen.