feat: add code delivered while on the exercise (#2)
* feat: display list of currencies using provided components

The instructions were to use the given developed components to display
the list with the header. Since no subcategories were needed,
SectionList was dismissed in favor of the most commonly used FlatList
which already supports Header rendering via props.

* fix: ui of given components due to bad layout on large scren

This fix comes from a fault of my own when setting the testbed
application and I didn't see that there was a misalignment issue on the
rendered list (because I was styling the FlatList when testing this but
the delivered code didn't have that styling).

* feat: filter by support of testmode

Initially, I thought about using a Pressable component to toggle states.
This would have been good in case we need to create our own toggling
component. However, as said by Iacopo, we didn't need that UI refinement
so I went to the ReactNative docs to pull the sample code for the Switch
component and use that instead. A few minor modifications were needed
since code was different but it was basically it.

Wanted to make a row view to put the text alongside it but during the
exercise it was expressed the lack of need of that so it was left as a
column container which doesn't make much sense visually but works for
the test.

* feat: sort by name and column

This step of the exercise was very troubling while working on it in the
heat of the moment. My first thought while doing it was to use the
`prevState` prop of the callback to avoid any cyclic depedencies with
`filteredData` (since it would be changed by the effect and then it
would retrigger it). However, I was biased by the continuous usage of
ESLint on the projects I worked were dependencies need to be put there,
always.

In the framework of this exercise, such check was not needed so the
easiest way to develop it was to change `filteredData`.

The approach I used was mutating the `prevState` element in place (by
the function sort). So this would not trigger a re-render on the page
which would resort in an unexpected UX.

The final workaround was to destruct the mutated element inside a new
array, effectively copying it and thus forcing a re-render of the page.

* fix: wrong reference (copy-paste issue) for useCurrencies import on tests

Copy-pasted code which was broken while developing the tests. Had fixed
the loading one but did not update the error one after that so it
continued to be broken.

* feat: add test to assert if filtering is working properly

The easiest way, although flawed, was to check the screen for any
`false` texts since they should have been filtered and not be rendered.

This is a bad test since it could happen that an element of the list has
either a name or code "false" so it would trigger a false negative.

Discussing with the team, they asked me if I was familiar with
`@testing-library` (to which I replied I used just a few times) because
there is a built-in function which is better suited for this.

After a quick google search while I was been told that it was no longer
needed I recalled that there's a queryByText function which would have
been better suited to test the lack of element in the screen.